l10/requirements.txt
numpy
scipy
torch
torchvision
pyro-ppl
matplotlib
seaborn
jupyter
tqdm
scikit-learn
